extract options interface instead of deriving
https://coveord.atlassian.net/browse/KIT-306
