+++ packages/headless/src/controllers/facets/_common/facet-option-definitions.ts
@@ -6,7 +6,7 @@ import {
   SchemaDefinition,
   StringValue,
 } from '@coveo/bueno';
-import {FacetSearchRequestOptions} from '../../../features/facets/facet-search-set/facet-search-request-options';
+import {FacetSearchRequestOptions} from '../../../api/search/facet-search/base/base-facet-search-request';
 
 /**
  * A unique identifier for the controller.
