+++ packages/headless/src/controllers/facets/category-facet/headless-category-facet-options.ts
@@ -1,10 +1,10 @@
 import {Schema, StringValue} from '@coveo/bueno';
+import {FacetSearchRequestOptions} from '../../../api/search/facet-search/base/base-facet-search-request';
 import {CategoryFacetRegistrationOptions} from '../../../features/facets/category-facet-set/interfaces/options';
 import {
   categoryFacetSortCriteria,
   CategoryFacetSortCriterion,
 } from '../../../features/facets/category-facet-set/interfaces/request';
-import {FacetSearchRequestOptions} from '../../../features/facets/facet-search-set/facet-search-request-options';
 import {
   facetId,
   field,
