+++ packages/headless/src/controllers/facets/facet/headless-facet-options.ts
@@ -1,5 +1,5 @@
 import {Schema, StringValue} from '@coveo/bueno';
-import {FacetSearchRequestOptions} from '../../../features/facets/facet-search-set/facet-search-request-options';
+import {FacetSearchRequestOptions} from '../../../api/search/facet-search/base/base-facet-search-request';
 import {FacetRegistrationOptions} from '../../../features/facets/facet-set/interfaces/options';
 import {facetSortCriteria} from '../../../features/facets/facet-set/interfaces/request';
 import {
